+++ src/__tests__/regular-macros.test.ts
@@ -3,15 +3,14 @@ import { parse } from "../parser/parser.js";
 import { regularMacrosAst, regularMacrosVoidFile } from "./fixtures/index.js";
 import assert from "node:assert";
 import path from "node:path";
-import { stdPath } from "../lib/parse-std.js";
 import { resolveFileModules } from "../modules.js";
-import { expandSyntaxMacrosOfFiles } from "../syntax-macros/index.js";
 import { expandRegularMacros } from "../regular-macros.js";
+import { stdPath } from "../parser/api/parse-std.js";
 
 describe("regular macro evaluation", () => {
   it("should evaluate macros of the example file", async () => {
     const parserOutput = parse(regularMacrosVoidFile);
-    const files = expandSyntaxMacrosOfFiles({ test: parserOutput });
+    const files = { test: parserOutput };
     const resolvedModules = resolveFileModules({
       files,
       srcPath: path.dirname("test"),
