+++ src/cli.ts
@@ -3,17 +3,15 @@ import { stdout } from "process";
 import { getConfig } from "./lib/config/index.js";
 import { genWasmCode } from "./wasm-code-gen.js";
 import { run } from "./run.js";
-import { parseDirectory, parseFile, parseStd, stdPath } from "./lib/index.js";
-import {
-  expandSyntaxMacrosOfFiles,
-  expandSyntaxMacros,
-} from "./syntax-macros/index.js";
 import { resolveFileModules } from "./modules.js";
 import { expandRegularMacros } from "./regular-macros.js";
 import { typeCheck } from "./semantics/index.js";
 import binaryen from "binaryen";
 import { resolveSrc } from "./lib/resolve-src.js";
 import { testGc } from "./lib/binaryen-gc/test.js";
+import { parseFile } from "./lib/index.js";
+import { parseDirectory } from "./parser/api/parse-directory.js";
+import { parseStd, stdPath } from "./parser/api/parse-std.js";
 
 main().catch(errorHandler);
 
@@ -64,8 +62,7 @@ async function getParserAst(index: string) {
 }
 
 async function getCoreAst(index: string) {
-  const parserAst = await getParserAst(index);
-  return expandSyntaxMacros(parserAst);
+  return await getParserAst(index);
 }
 
 async function getModuleAst(index: string) {
@@ -75,12 +72,10 @@ async function getModuleAst(index: string) {
     ? await parseDirectory(src.srcRootPath)
     : { [src.indexPath]: await parseFile(src.indexPath) };
 
-  const parsedFiles = {
+  const files = {
     ...srcFiles,
     ...(await parseStd()),
   };
-
-  const files = expandSyntaxMacrosOfFiles(parsedFiles);
 
   return resolveFileModules({
     files,
