+++ src/lib/index.ts
@@ -1,6 +1,4 @@
-export * from "./file.js";
-export * from "./token.js";
+export * from "../parser/char-stream.js";
+export * from "../parser/token.js";
 export * as grammar from "./grammar.js";
-export * from "./parse-directory.js";
-export * from "./parse-file.js";
-export * from "./parse-std.js";
+export * from "../parser/api/parse-file.js";
