+++ src/parser/api/parse-file.ts
@@ -1,0 +1,8 @@
+import fs from "node:fs";
+import { parse } from "../parser.js";
+import { List } from "../../syntax-objects/list.js";
+
+export const parseFile = async (path: string): Promise<List> => {
+  const file = fs.readFileSync(path, { encoding: "utf8" });
+  return parse(file);
+};
