+++ src/parser/syntax-macros/functional-notation.ts
