+++ src/parser/syntax-macros/primary.ts
