+++ src/parser/syntax-macros/types.ts
@@ -1,4 +1,4 @@
-import { List } from "../syntax-objects/index.js";
+import { List } from "../../syntax-objects/index.js";
 
 /** Takes the whole ast, returns a transformed version of the whole ast */
 export type SyntaxMacro = (list: List) => List;
