+++ src/semantics/README.md
@@ -1,0 +1,6 @@
+# Semantic Processing
+
+- Resolves Modules
+- Expands Regular (Function) Macros
+- Resolves Entities (Types, Functions, Variables, etc)
+- Checks Types
